requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
scikit-learn==1.5.2
pandas==2.2.3
numpy<2
mlflow==2.9.0
boto3==1.35.76
snowflake-connector-python==3.12.3
python-dotenv==1.0.1
prometheus-client==0.21.0
streamlit==1.40.2
plotly==5.24.1
shap==0.46.0
pytest==8.3.4
pytest-cov==6.0.0
httpx==0.28.1
evidently==0.4.40
pydantic==2.10.3
setuptools
